# Optimize iterator Vec/SetVec