```
goos: linux
goarch: amd64
cpu: AMD Ryzen 9 3950X 16-Core Processor

original: BenchmarkFloat32Iterator/Vec3
  float32Iterator-32  79926466   14.88 ns/op
  binaryIterator-32   26346526   45.54 ns/op

this PR: BenchmarkFloat32Iterator/Vec3
  float32Iterator-32  127507297  9.460 ns/op
  binaryIterator-32   26424267   45.57 ns/op
```